Fixed: 'Battery Percentage' output
I noticed the battery percentage reported by this integration is a little out. Around 9-10% in my case.

Looking at the code as compared with a CURL response, I noticed the CURL response provides the 'percentage_charged', whereas this integration doesn't use that, rather uses the min and max voltages to work out the percentage. I don't know if perhaps this additional property is new and that's why it's not used?